fix(pm): always delete in downstream services
Currently if a DELETE message is received but no matching PM is
found in the sync database the AfterDeleteEvent listener does not fire
and downstream services are not notified of the potential deletion.

In most cases this is fine, as the deletion has already occured, but in
edge cases where the DELETE message is required again (e.g. cleaning up
partial/leftover data from previously failed deletes) the message never
makes it to the downstream services leaving them out of date with no way
to automatically sync back up.

The original intention of the PM event listener was that it would
fire on database events, allowing async handling of updates and deletes
de-coupled from the incoming record messages. However, the
implementation uses application events instead, which offer no such
benefits. As a result the event listeners offer nothing other than
complexity/obscurity given the current needs.

Remove the delete handling from the ProgrammeMembershipEventListener and
update the ProgrammeMembershipSyncService to call the TCS sync directly.
This simplifies the application flow as well as removing the need for
cache trickery to keep data available after deletion.
Due to a circular dependency issue the "on save" aspects of the event
listener have not been removed at this time, as more significant
refactoring would be required than I'm willing to tackle as part of boy
scouting.

TIS21-6701

+++ src/main/java/uk/nhs/hee/tis/trainee/sync/service/ProgrammeMembershipSyncService.java
@@ -58,12 +58,14 @@ public class ProgrammeMembershipSyncService implements SyncService {
   private final ProgrammeMembershipMapper mapper;
   private final ApplicationEventPublisher eventPublisher;
 
+  private final TcsSyncService tcsService;
+
   ProgrammeMembershipSyncService(ProgrammeMembershipRepository repository,
       DataRequestService dataRequestService,
       FifoMessagingService fifoMessagingService,
       @Value("${application.aws.sqs.programme-membership}") String queueUrl,
       RequestCacheService requestCacheService, ProgrammeMembershipMapper mapper,
-      ApplicationEventPublisher eventPublisher) {
+      ApplicationEventPublisher eventPublisher, TcsSyncService tcsService) {
     this.repository = repository;
     this.dataRequestService = dataRequestService;
     this.queueUrl = queueUrl;
@@ -71,6 +73,7 @@ public class ProgrammeMembershipSyncService implements SyncService {
     this.mapper = mapper;
     this.eventPublisher = eventPublisher;
     this.fifoMessagingService = fifoMessagingService;
+    this.tcsService = tcsService;
   }
 
   @Override
@@ -94,8 +97,11 @@ public class ProgrammeMembershipSyncService implements SyncService {
     ProgrammeMembership programmeMembership = mapper.toEntity(programmeMembershipRecord.getData());
     UUID uuid = programmeMembership.getUuid();
 
+    boolean requested = false;
+
     if (programmeMembershipRecord.getOperation().equals(DELETE)) {
       repository.deleteById(uuid);
+      tcsService.syncRecord(programmeMembershipRecord);
     } else if (programmeMembershipRecord.getOperation().equals(LOOKUP)) {
       Optional<ProgrammeMembership> optionalProgrammeMembership = repository.findById(uuid);
 
@@ -105,12 +111,15 @@ public class ProgrammeMembershipSyncService implements SyncService {
         eventPublisher.publishEvent(event);
       } else {
         request(uuid);
+        requested = true;
       }
     } else {
       repository.save(programmeMembership);
     }
 
-    requestCacheService.deleteItemFromCache(ProgrammeMembership.ENTITY_NAME, uuid.toString());
+    if (!requested) {
+      requestCacheService.deleteItemFromCache(ProgrammeMembership.ENTITY_NAME, uuid.toString());
+    }
   }
 
   public Optional<ProgrammeMembership> findById(String uuid) {
